requirements/dev.txt
# mk42
# requirements/dev.txt

-r common.txt

pypiup==0.0.2
django-debug-toolbar==1.8
